+++ src/main/java/goojeans/harulog/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package goojeans.harulog.repository;
+package goojeans.harulog.category.repository;
 
-import goojeans.harulog.domain.entity.Category;
+import goojeans.harulog.category.domain.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
