+++ src/main/java/goojeans/harulog/challenge/domain/entity/Challenge.java
@@ -1,6 +1,8 @@
-package goojeans.harulog.domain.entity;
+package goojeans.harulog.challenge.domain.entity;
 
+import goojeans.harulog.category.domain.entity.Category;
 import goojeans.harulog.chat.domain.entity.ChatRoom;
+import goojeans.harulog.domain.entity.BaseEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
