+++ src/main/java/goojeans/harulog/challenge/util/ChallengeRole.java
@@ -1,4 +1,4 @@
-package goojeans.harulog.util;
+package goojeans.harulog.challenge.util;
 
 public enum ChallengeRole {
     LEADER, PARTICIPANT
