+++ src/test/java/goojeans/harulog/category/repository/CategoryRepositoryTest.java
@@ -1,6 +1,6 @@
-package goojeans.harulog.repository;
+package goojeans.harulog.category.repository;
 
-import goojeans.harulog.domain.entity.Category;
+import goojeans.harulog.category.domain.entity.Category;
 import lombok.extern.slf4j.Slf4j;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 @Slf4j
 @DataJpaTest
